Make do_stats more manageable
